+++ CHANGELOG.md
@@ -30,6 +30,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fixed ARCO data source `ARCO_TIME_STOP` fallback to 2025-12-31,
+  reflecting the most recent available data in the bucket
+
 ### Security
 
 ### Dependencies